refactor: Separate Header.module.scss from Home.module.scss
