+++ components/header/Header.module.scss
@@ -1,0 +1,33 @@
+@import '@/styles/variables';
+
+.title {
+  margin: 0;
+  font-size: 4rem;
+  text-align: center;
+
+  a {
+    color: $primary-purple;
+    text-decoration: none;
+
+    &:hover,
+    &:focus,
+    &:active {
+      text-decoration: underline;
+    }
+  }
+}
+
+.description {
+  font-size: 1.5rem;
+  text-align: center;
+}
+
+@media (max-width: 600px) {
+  .title {
+    font-size: 2rem;
+  }
+
+  .description {
+    font-size: 1rem;
+  }
+}
