+++ pages/index.tsx
@@ -6,8 +6,8 @@ import utc from 'dayjs/plugin/utc';
 import { GetStaticProps } from 'next';
 import Head from 'next/head';
 import Footer from '@/components/Footer';
-import Header from '@/components/Header';
 import Card from '@/components/card/Card';
+import Header from '@/components/header/Header';
 import { Pagination } from '@/components/pagination/Pagination';
 import { microcms } from '@/libs/microcms';
 import styles from '@/styles/Home.module.scss';
