+++ styles/Home.module.scss
@@ -41,28 +41,6 @@
   }
 }
 
-.title {
-  margin: 0;
-  font-size: 4rem;
-  text-align: center;
-
-  a {
-    color: $primary-purple;
-    text-decoration: none;
-
-    &:hover,
-    &:focus,
-    &:active {
-      text-decoration: underline;
-    }
-  }
-}
-
-.description {
-  font-size: 1.5rem;
-  text-align: center;
-}
-
 .grid {
   display: flex;
   flex-wrap: wrap;
